# server/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "ACS Energy CRM API",
        "description": "Lead cockpit for EV rollouts — CHS, Hotels, Corporate Parks, and Developers across MMR.",
        "version": "1.0.0",
        "contact": {
            "name": "ACS Energy"
        }
    },
    "servers": [
        {
            "url": "https://server-nine-olive-75.vercel.app",
            "description": "Production"
        },
        {
            "url": "http://localhost:5000",
            "description": "Local development"
        }
    ],
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
...
    },
    "paths": {
...
        "/api/users": {
            "get": {
                "summary": "List all users (admin only)",
                "tags": [
                    "Users"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of User objects"
                    }
                }
            },
            "post": {
                "summary": "Create a new user (admin only)",
                "tags": [
                    "Users"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "admin",
                                            "sales"
                                        ],
                                        "default": "sales"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created"
                    },
                    "400": {
                        "description": "Missing required fields"
                    },
                    "409": {
                        "description": "Email already exists"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "delete": {
                "summary": "Delete a user (admin only)",
                "tags": [
                    "Users"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User removed"
                    },
                    "400": {
                        "description": "Cannot delete self or last admin"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        }
    }
}